restructure match_ methods
