+++ src/lib.rs
@@ -1,11 +1,12 @@
 #![allow(clippy::type_complexity)]
 
 pub use tree::{
-    InMemNodeId, InMemStorage, SizedStrideNode, SortableNodeId, StorageBackend, Stride, TreeBitMap,
-    TreeBitMapNode, CacheGuard, PrefixCacheGuard
+    CacheGuard, InMemNodeId, InMemStorage, PrefixCacheGuard, SizedStrideNode, SortableNodeId,
+    StorageBackend, Stride, TreeBitMap, TreeBitMapNode,
 };
 
 pub mod common;
+mod query;
 pub mod synth_int;
 pub mod tree;
